services/query-api-service/requirements.txt
# ─── Core ──────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
redis[asyncio]==5.0.4
aio-pika==9.4.1           # RabbitMQ async consumer (persist events)

# ─── Validación ────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ─── Auth (validar JWT) ────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ─── Observabilidad ────────────────────────────────────────────────────────
structlog==24.2.0
python-json-logger==2.0.7

# ─── Utils ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==8.3.0
